chaos-app/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
httpx==0.27.0
psycopg2-binary==2.9.9
psutil==5.9.8
pyyaml==6.0.1
python-dotenv==1.0.1
azure-identity==1.16.0
azure-monitor-ingestion==1.0.3
